Move MatchJob dataclass from grouping.py to config.py
MatchJob is a plain data container that belongs with the other
config/data types, not in the grouping module.
